move dumpState to its own file
